Extract the rest of the boilerplate in `rules.rs` to the proc macro
