+++ crates/oxc_linter/src/rules.rs
@@ -5,67 +5,3 @@ oxc_macros::declare_all_lint_rules! {
     no_empty_pattern,
     deepscan::uninvoked_array_callback,
 }
-
-use crate::{context::LintContext, rule::Rule, rule::RuleMeta, AstNode};
-
-lazy_static::lazy_static! {
-    pub static ref RULES: Vec<RuleEnum> = vec![
-        RuleEnum::NoDebugger(NoDebugger::default()),
-        RuleEnum::NoEmpty(NoEmpty::default()),
-        RuleEnum::NoArrayConstructor(NoArrayConstructor::default()),
-        RuleEnum::NoEmptyPattern(NoEmptyPattern::default()),
-        RuleEnum::UninvokedArrayCallback(UninvokedArrayCallback::default()),
-    ];
-}
-
-#[derive(Debug, Clone)]
-#[allow(clippy::enum_variant_names)]
-pub enum RuleEnum {
-    NoDebugger(NoDebugger),
-    NoEmpty(NoEmpty),
-    NoArrayConstructor(NoArrayConstructor),
-    NoEmptyPattern(NoEmptyPattern),
-    UninvokedArrayCallback(UninvokedArrayCallback),
-}
-
-impl RuleEnum {
-    pub const fn name(&self) -> &'static str {
-        match self {
-            Self::NoDebugger(_) => NoDebugger::NAME,
-            Self::NoEmpty(_) => NoEmpty::NAME,
-            Self::NoArrayConstructor(_) => NoArrayConstructor::NAME,
-            Self::NoEmptyPattern(_) => NoEmptyPattern::NAME,
-            Self::UninvokedArrayCallback(_) => UninvokedArrayCallback::NAME,
-        }
-    }
-
-    pub fn read_json(&self, maybe_value: Option<serde_json::Value>) -> Self {
-        match self {
-            Self::NoDebugger(_) => Self::NoDebugger(
-                maybe_value.map(NoDebugger::from_configuration).unwrap_or_default(),
-            ),
-            Self::NoEmpty(_) => {
-                Self::NoEmpty(maybe_value.map(NoEmpty::from_configuration).unwrap_or_default())
-            }
-            Self::NoArrayConstructor(_) => Self::NoArrayConstructor(
-                maybe_value.map(NoArrayConstructor::from_configuration).unwrap_or_default(),
-            ),
-            Self::NoEmptyPattern(_) => Self::NoEmptyPattern(
-                maybe_value.map(NoEmptyPattern::from_configuration).unwrap_or_default(),
-            ),
-            Self::UninvokedArrayCallback(_) => Self::UninvokedArrayCallback(
-                maybe_value.map(UninvokedArrayCallback::from_configuration).unwrap_or_default(),
-            ),
-        }
-    }
-
-    pub fn run<'a>(&self, node: &AstNode<'a>, ctx: &LintContext<'a>) {
-        match self {
-            Self::NoDebugger(rule) => rule.run(node, ctx),
-            Self::NoEmpty(rule) => rule.run(node, ctx),
-            Self::NoArrayConstructor(rule) => rule.run(node, ctx),
-            Self::NoEmptyPattern(rule) => rule.run(node, ctx),
-            Self::UninvokedArrayCallback(rule) => rule.run(node, ctx),
-        }
-    }
-}
